% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fn.R
\name{base_heatmap}
\alias{base_heatmap}
\title{Base R cluster heatmap}
\usage{
base_heatmap(cl_col_dat2, col_specs, res_num, chromo)
}
\arguments{
\item{cl_col_dat2}{Tibble collecting all heatmap entries to plot with corresponding coordinates}

\item{col_specs}{List collecting the necessary components to build color-map}

\item{res_num}{Named vector with the resolutions and their labels present in the HiC data}

\item{chromo}{String indicating the chromosome containing the considered cluster}
}
\value{
Heatmap plot produced by the plotgardener backend
}
\description{
Base R cluster heatmap
}
