% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\docType{data}
\name{chr_dat_l}
\alias{chr_dat_l}
\title{Trial data for trying BHiCect 2.0}
\format{
\subsection{\code{chr_dat_l}}{

A list of tibbles with each elements collecting HiC data for chromosome 22 at a particular resolution for the HMEC cell line from Rao et al. 2014 :
\describe{
\item{X1}{HiC bin as integer}
\item{X2}{HiC bin as integer}
\item{X3}{KR normalised HiC signal}
\item{weight}{Box Cox transformed HiC signal}

...
}
}
}
\usage{
chr_dat_l
}
\description{
Trial data for trying BHiCect 2.0
}
\keyword{datasets}
