% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fn.R
\name{custom_color_map_fn}
\alias{custom_color_map_fn}
\title{Utility function to produce color maps}
\usage{
custom_color_map_fn(res_set, palette)
}
\arguments{
\item{res_set}{Set of resolution over which to map colors}

\item{palette}{Palette to use for color map}
}
\value{
list with color map hexdecimal values and corresponding breaks
}
\description{
Utility function to produce color maps
}
