% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{extract_bpt_cl_dat_fn}
\alias{extract_bpt_cl_dat_fn}
\title{Extract HiC data for full target cluster genealogy}
\usage{
extract_bpt_cl_dat_fn(cl_lvl_tbl, cl_dat_l, res_num, tmp_cl_res_set, nworkers)
}
\arguments{
\item{cl_lvl_tbl}{Summary table of target cluster genealogy}

\item{cl_dat_l}{Subset of HiC data for target cluster}

\item{res_num}{Named vector collecting observed resolution for chromosome-wide HiC data}

\item{tmp_cl_res_set}{Subset of HiC data resolution observed in target cluster}

\item{nworkers}{Number of workers for parallel computation}
}
\value{
Tibble with one row per cluster and a column listing corresponding HiC data at matching resolution and interpolated high-resolution
}
\description{
Extract HiC data for full target cluster genealogy
}
