% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{ff}
\alias{ff}
\title{Title}
\usage{
ff(x)
}
\arguments{
\item{x}{list for which we wish to output all the nested lists within it}
}
\value{
a character vector tracing the full path of the nested lists
}
\description{
Title
}
