% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{find_genomic_runs}
\alias{find_genomic_runs}
\title{Find Contiguous Genomic Intervals}
\usage{
find_genomic_runs(ids, res_level)
}
\arguments{
\item{ids}{A numeric or integer vector of starting coordinate positions.}

\item{res_level}{Integer. The base-pair resolution tier size (e.g., \code{10000})}
}
\value{
A \code{list} of items containing \code{start} and \code{end} values
of contiguous fragments, or \code{NULL} if \code{ids} is empty.
}
\description{
Evaluates a vector of sparse genomic bin positions at a specified resolution
and compresses them into a list of continuous genomic interval blocks.
}
