% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fn.R
\name{full_f_mat}
\alias{full_f_mat}
\title{Build heatmap matrix}
\usage{
full_f_mat(cl_mat, res, val_col)
}
\arguments{
\item{cl_mat}{A data frame containing an edge-list of clusters. This means that each row should represent a connection between two elements (called \code{ego} and \code{alter}), along with their associated values.}

\item{res}{A numeric value representing the resolution of the heatmap. This dictates the size of the bins used for grouping the data.}

\item{val_col}{A string indicating the name of the column in \code{cl_mat} that contains the values to be used for coloring the heatmap.}
}
\value{
A sparse matrix where each entry corresponds to a cluster relationship, designed for heatmap visualization. The matrix will have cluster values filled in based on the provided edge-list.
}
\description{
This function creates a heatmap matrix from cluster data, which can be used for visualizing relationships or interactions between different elements in a dataset.
}
