% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{get_obj_in_fn}
\alias{get_obj_in_fn}
\title{Import R objects}
\usage{
get_obj_in_fn(file)
}
\arguments{
\item{file}{string containing the path to the R object to import}
}
\value{
R objects
}
\description{
Import R objects
}
