% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy_get_adj_mat_fn}
\alias{legacy_get_adj_mat_fn}
\title{Produce adjacency matrix from tibble}
\usage{
legacy_get_adj_mat_fn(g_chr1)
}
\arguments{
\item{g_chr1}{Igraph object from which to extract appropriate adjacency matrix}
}
\value{
The square and symmetric Adjacency matrix from the input graph
}
\description{
Produce adjacency matrix from tibble
}
