% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy_lp_fn}
\alias{legacy_lp_fn}
\title{Laplacian function}
\usage{
legacy_lp_fn(x)
}
\arguments{
\item{x}{Square and symmetric matrix}
}
\value{
list of two elements:
\itemize{
\item Tibble with two columns, one for each eigen-vector associated with the two smallest eigen-values
\item Vector of the two smallest eigen-values
}
}
\description{
Function computing the Laplacian
}
