% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy_select_best_res_fn}
\alias{legacy_select_best_res_fn}
\title{Select Best resoluion function}
\usage{
legacy_select_best_res_fn(
  i,
  tmp_chr_dat_l,
  chr1_tree_cl,
  res_num,
  tmp_res,
  tmp_res_set,
  min_res
)
}
\arguments{
\item{i}{Considered cluster}

\item{tmp_chr_dat_l}{List of tibble containing the cluster HiC at multiple resolutions}

\item{chr1_tree_cl}{List with the bin content of the already found clusters}

\item{res_num}{Named numeric vector with the HiC resolutions present in the data}

\item{tmp_res}{Current cluster resolution}

\item{tmp_res_set}{Named numeric vector containing the subset of HiC resolution applicable for the considered cluster}

\item{min_res}{Character indicating the minimum bin-size available for HiC data}
}
\value{
Character with the best resolution at which to perform bi-partition
}
\description{
function selecting the best resolution at which to perform clustering
}
