% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy_simple_partition_tbl_fn}
\alias{legacy_simple_partition_tbl_fn}
\title{Simple partition table function}
\usage{
legacy_simple_partition_tbl_fn(lp_res, tmp_res)
}
\arguments{
\item{lp_res}{Output from the Laplacian function \link{lp_fn}}

\item{tmp_res}{Resolution at which to perform bi-partition}
}
\value{
Tibble with the partition statistics and ensuing clusters
}
\description{
Function producing the statistics enabling bi-partition
}
