% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy_ss}
\alias{legacy_ss}
\title{Sum of square function}
\usage{
legacy_ss(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sum of square error
}
\description{
Utility function to compute sum of squared errors
}
