% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{part_cond_calc}
\alias{part_cond_calc}
\title{Bipartition function}
\usage{
part_cond_calc(x, reff_g, tmp_res)
}
\arguments{
\item{x}{Output eigen-decomposition step}

\item{reff_g}{Original graph on which we computed the eigen decomposition}

\item{tmp_res}{The resolution at which we are performing spectral clustering}
}
\value{
Returns a list of two elements, corresponding to the bi-partition of the input graph
}
\description{
Delineate bi-partitions using kmeans
}
