% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{partition_fn}
\alias{partition_fn}
\title{Title}
\usage{
partition_fn(reff_g, smpl_thresh_tbl, tmp_res, cl_var)
}
\arguments{
\item{reff_g}{Parent cluster igraph network object}

\item{smpl_thresh_tbl}{Tibble with bi-partition statistics}

\item{tmp_res}{Resolution at which cluster is partitioned}

\item{cl_var}{Variable from smpl_thresh_tbl to use for partition}
}
\value{
list where each element is the bin content of the produced paritions
}
\description{
Title
}
