% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt_viz_utils.R
\name{produce_bpt_cl_lvl_tbl}
\alias{produce_bpt_cl_lvl_tbl}
\title{Produce summary table of cluster genealogy}
\usage{
produce_bpt_cl_lvl_tbl(chr_spec_res, tmp_cl)
}
\arguments{
\item{chr_spec_res}{BHiCect result object}

\item{tmp_cl}{Target cluster name}
}
\value{
Table summarising the nested clusters within the target cluster
}
\description{
Produce summary table of cluster genealogy
}
