% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{produce_bpt_heat_tbl}
\alias{produce_bpt_heat_tbl}
\title{Wrapper function to produce edgelist for heatmap visualisation with BPT segmentation}
\usage{
produce_bpt_heat_tbl(
  tmp_cl,
  chr_spec_res,
  tmp_cl_res_set,
  res_num,
  chr_dat_l,
  nworkers
)
}
\arguments{
\item{tmp_cl}{Target cluster}

\item{chr_spec_res}{BHiCect result object}

\item{tmp_cl_res_set}{Named vector for HiC resolution observed within target cluster}

\item{res_num}{Named vector collecting observed resolution for chromosome-wide HiC data}

\item{chr_dat_l}{Chromosome-wide HiC data for target cluster}

\item{nworkers}{Number of workers for parallel computation}
}
\value{
3-column edge-list tibble with edge-weight corresponding to color-map values
}
\description{
Wrapper function to produce edgelist for heatmap visualisation with BPT segmentation
}
