% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mres_viz_utils.R
\name{produce_mres_color_tbl}
\alias{produce_mres_color_tbl}
\title{Convert HiC to colormap}
\usage{
produce_mres_color_tbl(cl_hires_dat_l, tmp_res_set)
}
\arguments{
\item{cl_hires_dat_l}{List of HiC data at observed resolution for target cluster interpolated to highest resolution}

\item{tmp_res_set}{Set of resolution at which target cluster exists}
}
\value{
tibble with HiC data converted to colormap values
}
\description{
Convert HiC to colormap
}
