% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mres_viz_utils.R
\name{produce_mres_hires_interpolation}
\alias{produce_mres_hires_interpolation}
\title{Interpolate HiC data to highest resolution}
\usage{
produce_mres_hires_interpolation(tmp_res_set, cl_dat_l, res_num, nworkers)
}
\arguments{
\item{tmp_res_set}{Target cluster set of resolutions}

\item{cl_dat_l}{List where each element correspond to a tibble with HiC data for the target cluster at particular resolution}

\item{res_num}{Named vector collecting observed resolution for chromosome-wide HiC data}

\item{nworkers}{Number of workers used for parallelised computation}
}
\value{
List of tibbles with corresponding HiC data converted to highest resolution
}
\description{
Interpolate HiC data to highest resolution
}
