% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_obj}
\alias{res_obj}
\title{Toy Resolution Mapping Result Object}
\format{
A list containing multiple structural slots:
\describe{
\item{nodes}{A tibble tracking multi-tier cluster metrics (id, res_level, depth, perf, obs)}
\item{edges}{A tibble reporting all the edges to build the BHiCect clustering tree}
}
}
\source{
Pre-computed test run from the BHiCect2 pipeline.
}
\usage{
res_obj
}
\description{
A sample output object containing hierarchical partitioning paths, node
architectures, and structural data frames for vignette demonstrations.
}
\keyword{datasets}
