% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{BHiCect}
\alias{BHiCect}
\title{BHiCect 2.0}
\usage{
BHiCect(res_set, res_num, chr_dat_l, cl_var = "smpl.cl", nworkers)
}
\arguments{
\item{res_set}{Character vector with the HiC resolutions present in the data}

\item{res_num}{Named numeric vector with the HiC resolutions present in the data}

\item{chr_dat_l}{List of tibble containing the chromosome-wide HiC at multiple resolutions}

\item{cl_var}{Variable from smpl_thresh_tbl to use for partition}

\item{nworkers}{Number of processes to launch for parallel computation}
}
\value{
List with:
\itemize{
\item Named list collecting the bin content of each found cluster
\item Edge-list for the bi-partition tree
\item Partition statistic table for each cluster
}
}
\description{
Main clustering function implementing recursive spectral clustering of HiC data
}
