% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{BHiCect}
\alias{BHiCect}
\title{BHiCect}
\usage{
BHiCect(chr1_mat, g_chr1, res_set, res_num, chr_dat_l)
}
\arguments{
\item{chr1_mat}{Hi-C data square matrix representation}

\item{g_chr1}{Hi-C data igraph representation}

\item{res_set}{Character vector of the available resolutions}

\item{res_num}{Named numeric vector of the available resolutions}

\item{chr_dat_l}{List with the HiC data at the available resolution}
}
\value{
a list with two elements, one containing a nested list representation of the clustering tree and another with a named list containing the bin content of each cluster found
}
\description{
Main function running the BHiCect algorithm
}
