% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{get_adj_mat_fn}
\alias{get_adj_mat_fn}
\title{Produce adjacency matrix from tibble}
\usage{
get_adj_mat_fn(g_chr1)
}
\arguments{
\item{g_chr1}{Igraph object from which to extract appropriate adjacency matrix}
}
\value{
The square and symmetric Adjacency matrix from the input graph
}
\description{
Produce adjacency matrix from tibble
}
