% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHiCect.R
\name{part_cond_calc}
\alias{part_cond_calc}
\title{Title}
\usage{
part_cond_calc(x, reff_g, tmp_res)
}
\arguments{
\item{x}{is the eigen-decomposition step output}

\item{reff_g}{is the original graph on which we computed the eigen decomposition}

\item{tmp_res}{the resolution at which we are performing spectral clustering}
}
\description{
Title
}
