% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{produce_bpt_cl_dat_l}
\alias{produce_bpt_cl_dat_l}
\title{Subset chromosome-wide HiC data to focus on target cluster}
\usage{
produce_bpt_cl_dat_l(tmp_cl_res_set, cl_lvl_tbl, chr_dat_l)
}
\arguments{
\item{tmp_cl_res_set}{Subset of resolutions observed in target cluster}

\item{cl_lvl_tbl}{Summary table of target cluster genealogy}

\item{chr_dat_l}{Chromosome-wide HiC data for target cluster}
}
\value{
List of HiC data with one element per resolution
}
\description{
Subset chromosome-wide HiC data to focus on target cluster
}
