% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{produce_bpt_color_tbl}
\alias{produce_bpt_color_tbl}
\title{Produce edgelist for heatmap visualisation based on BPT segmentation}
\usage{
produce_bpt_color_tbl(cl_lvl_tbl)
}
\arguments{
\item{cl_lvl_tbl}{Summary table for target cluster genealogy}
}
\value{
Edgelist with colormap value for correponding resolutions
}
\description{
Produce edgelist for heatmap visualisation based on BPT segmentation
}
